requirements.txt
streamlit
openai-agents
pydantic
typing-extensions